Update structure in simple table
